new ui search empty with minor refactoring
