+++ public/locales/en-US/common.json
@@ -2,7 +2,7 @@
   "page": "page",
   "goTo": "Go to",
   "search": "search",
-  "back": "back",
+  "back": "Go Back",
   "notFoundMessage": "Your input \"{{search}}\" did not match any record.\n\nPlease make sure your input contains only one of the following items:",
   "pageNotFound": "404 | This page could not be found",
   "prev": "Prev",
